requirements.txt
fastapi
uvicorn
torch
transformers
sentencepiece
peft
boto3
python-dotenv
bitsandbytes
accelerate